backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic[email]==2.7.0
email-validator==2.1.0
python-dotenv==1.0.0
supabase==2.7.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
google-generativeai==0.8.3
pytest==7.4.3
pytest-asyncio==0.21.1
anyio[trio]==4.3.0
